Day8

+++ src/main/java/com/sinkovits/aoc2023/AdventOfCodeDailyExercise.java
@@ -2,6 +2,6 @@ package com.sinkovits.aoc2023;
 
 public interface AdventOfCodeDailyExercise {
 
-    int solveFirst();
-    int solveSecond();
+    long solveFirst();
+    long solveSecond();
 }
