+++ src/main/java/com/sinkovits/aoc2023/Day3.java
@@ -125,7 +125,7 @@ public class Day3 implements AdventOfCodeDailyExercise {
     }
 
     @Override
-    public int solveFirst() {
+    public long solveFirst() {
         LineProcessor<Day3Context> lineProcessor = getDay3ContextLineReader();
         Day3Context context = lineProcessor.processLines(this::processLineSolution1);
         log.info("Solution for the first exercise: {}", context.getSum());
@@ -133,7 +133,7 @@ public class Day3 implements AdventOfCodeDailyExercise {
     }
 
     @Override
-    public int solveSecond() {
+    public long solveSecond() {
         LineProcessor<Day3Context> lineProcessor = getDay3ContextLineReader();
         Day3Context context = lineProcessor.processLines(this::processLineSolution2);
         log.info("Solution for the second exercise: {}", context.getSum());
