+++ src/main/java/com/sinkovits/aoc2023/Day6.java
@@ -29,7 +29,7 @@ public class Day6 implements AdventOfCodeDailyExercise {
     }
 
     @Override
-    public int solveFirst() {
+    public long solveFirst() {
         LineProcessor<Context> lineProcessor = getContextLineReader();
         Context context = lineProcessor.processLines(this::processLineSolution1);
         long result = calculateResult(context);
@@ -38,7 +38,7 @@ public class Day6 implements AdventOfCodeDailyExercise {
     }
 
     @Override
-    public int solveSecond() {
+    public long solveSecond() {
         LineProcessor<Context> lineProcessor = getContextLineReader();
         Context context = lineProcessor.processLines(this::processLineSolution2);
         long result = calculateResult(context);
